requirements.txt
asn1crypto==0.22.0
bcrypt==3.1.3
blessings==1.6
bpython==0.16
certifi==2017.4.17
cffi==1.10.0
chardet==3.0.4
cryptography==1.9
curtsies==0.2.11
dj-database-url==0.4.2
Django==1.8.18
django-admin-bootstrapped==2.5.7
django-allauth==0.32.0
django-braces==1.11.0
django-configurations==2.0
django-datetime-widget==0.9.3
django-debug-toolbar==1.8
django-materializecss-form==1.0.1
django-mptt==0.8.7
django-mptt-admin==0.4.6
django-versatileimagefield==1.7.1
django-widget-tweaks==1.4.1
djangorestframework==3.6.3
enum34==1.1.6
Fabric==1.13.2
fabtools==0.20.0
greenlet==0.4.12
gunicorn==19.7.1
hiredis==0.2.0
idna==2.5
ipaddress==1.0.18
oauthlib==2.0.2
olefile==0.44
paramiko==2.2.1
Pillow==4.0.0
psycopg2==2.7.1
pyasn1==0.2.3
pycparser==2.17
Pygments==2.2.0
PyNaCl==1.1.2
python-http-client==2.3.0
python-openid==2.2.5
pytz==2017.2
requests==2.18.1
requests-oauthlib==0.8.0
sendgrid==3.6.5
sendgrid-django==4.0.4
six==1.10.0
sqlparse==0.2.3
urllib3==1.21.1
wcwidth==0.1.7
